+++ lib/remote-api.js
@@ -1,48 +1,3 @@
-var superagent = require('superagent');
+var remoteApiImpl = require('./remote-api-impl');
 
-module.exports = {
-
-  mock: function(opts, callback) {
-    superagent.post("http://localhost:7001/syncMock")
-      .type("json")
-      .send(opts)
-      .end(function(err, res) {
-        if (callback) callback(err, res)
-      });
-  },
-
-  defaults: function(opts, callback) {
-    superagent.post("http://localhost:7001/syncDefaults")
-      .type("json")
-      .send(opts)
-      .end(function(err, res) {
-        if (callback) callback(err, res)
-      });
-  },
-
-  lastRequest: function(method, path, callback) {
-    superagent.get("http://localhost:7001/lastRequest")
-      .set('method', method)
-      .set('path', path)
-      .end(function(err, res) {
-        if (callback) callback(err, res)
-      });
-  },
-
-  lastRequests: function(method, path, callback) {
-    superagent.get("http://localhost:7001/lastRequests")
-      .set('method', method)
-      .set('path', path)
-      .end(function(err, res) {
-        if (callback) callback(err, res)
-      });
-  },
-
-  clearLastRequests: function(callback) {
-    superagent.delete("http://localhost:7001/clearLastRequests")
-      .end(function(err, res) {
-        if (callback) callback(err, res)
-      });
-  }
-
-};
+module.exports = remoteApiImpl({ baseUrl: "http://localhost:7001" });
